Update diagd to load configs generated by go-kubewatch.
